requirements.txt
python-dotenv == 0.18.0
aiogram == 2.14.3
aiofiles==0.7.0
aiohttp==3.8.1
beautifulsoup4==4.9.3
emoji==1.7.0